# backend/api/openapi.yml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
  description: API documentation for the application.
servers:
  - url: http://localhost:3002
    description: Local dev server
paths:
...
  /admin:
    post:
      summary: Admin
      operationId: loginAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRequest"
      responses:
        "200":
          description: Login exitoso admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminResponse"
        "401":
          description: Credenciales inválidas
        "500":
          description: Error en el servidor
      x-eov-operation-handler: admin
  /login:
    post:
      summary: Login
      operationId: loginUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Credenciales inválidas
        "500":
          description: Error en el servidor
      x-eov-operation-handler: usuario
  /usuario:
    get:
      summary: Obtener usuarios
      operationId: obtenerUsuarios
      responses:
        "200":
          description: Usuarios obtenidos correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UsuariosResponse"
        "500":
          description: Error en el servidor
      x-eov-operation-handler: usuario
    post:
      summary: Usuario
      operationId: addUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioRequest"
      responses:
        "201":
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioResponse"
        "400":
          description: Error en la solicitud
        "500":
          description: Error en el servidor
      x-eov-operation-handler: usuario
...
components:
  schemas:
...
    UsuariosResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Juan Perez"
        correo:
          type: string
          example: "juan@gmail.com"
        nickname:
          type: string
          example: "jperez"
        creado_en:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"
...
    AdminRequest:
      type: object
      properties:
        nickname:
          type: string
          example: "admin"
        contrasena:
          type: string
          example: "admin123"
    AdminResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Admin"
        nickname:
          type: string
          example: "admin"
        contrasena:
          type: string
          example: "admin123"
        creado_en:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"
    LoginRequest:
      type: object
      properties:
        nickname:
          type: string
          example: "jperez"
        contrasena:
          type: string
          example: "password123"
    LoginResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Juan Perez"
        nickname:
          type: string
          example: "jperez"
...
    UsuarioRequest:
      type: object
      properties:
        nombre:
          type: string
          example: "Juan Perez"
        correo:
          type: string
          example: "juan@gmail.com"
        nickname:
          type: string
          example: "jperez"
        contrasena:
          type: string
          example: "password123"
    UsuarioResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Juan Perez"
        correo:
          type: string
          example: "juan@gmail.com"
        nickname:
          type: string
          example: "jperez"
        creado_en:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"